signtalk_web/signtalk/requirements.txt
# SignTalk 手语识别视频通话系统依赖包（兼容无GPU+Python3.10）

# ==================== Django相关 ====================
Django==4.2.17
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
PyMySQL==1.1.0

# ==================== 手语识别核心依赖 ====================
# MediaPipe手语识别
mediapipe==0.10.11

# 计算机视觉
opencv-python==4.10.0.84
opencv-contrib-python==4.10.0.84

# 深度学习框架（CPU版，兼容无GPU）
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

# TensorFlow (CPU版，可选CNN模型)
tensorflow-intel==2.13.0
tensorflow-estimator==2.13.0

# ==================== 数据处理 ====================
numpy==1.24.3  # 适配TensorFlow 2.13.0的<=1.24.3要求
scipy==1.10.1
pandas==2.0.3

# 图像处理
Pillow==10.4.0

# ==================== 机器学习支持库 ====================
# JAX生态系统 (MediaPipe依赖)
jax==0.4.13
jaxlib==0.4.13
ml-dtypes==0.2.0

# 数学和科学计算
sympy==1.13.3
mpmath==1.3.0

# ==================== 网络和通信 ====================
requests==2.31.0
urllib3==2.0.7

# ==================== 系统工具 ====================
# 文件处理
filelock==3.16.1
fsspec==2024.12.0

# 类型支持
typing-extensions==4.5.0  # 同时兼容PyTorch 2.0.1和TensorFlow 2.13.0

# 其他工具
absl-py==2.1.0
attrs==24.3.0
cffi==1.17.1
pycparser==2.22

# ==================== 可视化和绘图 ====================
matplotlib==3.7.5
contourpy==1.1.1
cycler==0.12.1
fonttools==4.55.3
kiwisolver==1.4.7
packaging==24.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
six==1.17.0

# ==================== 网络和序列化 ====================
# Protocol Buffers
protobuf==3.20.3
flatbuffers==24.12.23

# 网络相关
networkx==3.1
opt-einsum==3.4.0

# ==================== 音频处理 ====================
sounddevice==0.5.1

# ==================== 模板和渲染 ====================
Jinja2==3.1.5
MarkupSafe==2.1.5

# ==================== 元数据和资源 ====================
importlib-metadata==8.5.0
importlib-resources==6.4.5
